Rework RomFS class
This class should now be used as the centralized object that takes care of all File IO with the romfs folder.

When accessing files, all other classes will go through this API for its romfs needs.

The old system saw lots of spaghetti throughout all classes interacting with romfs in their own ways. This will centralize that.

If I could make this class into a singleton I would, as that is the idea of the class. For now it will just be a single reference passed around the classes. Honour system.