Appease warnings
There were like 15+ warnings in VS, I managed to clean up the code and get them down to only 4 that won't be fixed until some more of the project is done.

This involved making various nullable fields nullable as they should be.

Also changed a bit of logic with Course.GetArea returning null on error. Now it will throw on error, and the caller will do error checking in a try/catch block rather then checking if the output is null.

Changed some logic in DoAreaParams(). The logic is still equivalent, but restructured in a way that appeases VS warnings.