Move DoAreaScene to new AreaScene class
DoAreaScene now broken down into multiple more modular functions inside of a class.
Create a new class by supplying a CourseArea in the constructor and call the DisplayArea method to create a canvas with the level drawn on it.

This should clean up Program.cs considerably and make editing the canvas display algorithms easier.