Fix area switch viewport bug
Upon the switch to using the LevelViewport class, there was a bug where the viewport field in CourseScene was not being updated on a tab switch. This meant that the viewport would stay the same, but the other panels would switch to reflect the area changes.

Now the viewport is updated on area switch, and all is well.