Clean up CourseScene
Removed unused fields that are now being taken over by LevelViewport.

Removed mParentWindow field and removed it from constructor. My apologies if someone was going to use this, but it wasn't being used for anything.

Added a '?' in Course.cs to appease an editor warning. This isn't particularly important but I thought I would keep it in this cleaning commit.