Change thumbnail loading
Thumbnail loading now done on RomFS path set, since it can be done in the background. This improves performance on the course select, since it is likely that the thumbnails will be loaded by the time the user clicks on it anyways, or there will just be black rectangles before the thumbnails load in. Before there was a bit of a hang upon selecting each new world.