Fix romfs filepath change crash
Fix crash that could happen if the user changed the romfs filepath outside of the program.

This crash was caused by the program attempting to open the latestCourse despite RomFS not being valid.

This was fixed by only letting the paramDB and latestCourse load if the romfs path is valid.